Add tests on the experiment type
